+++ cmd/network_validator_deposit_alt.go
@@ -21,11 +21,7 @@ address. Remember it will need your wallet address and private keys. Thus it wil
 This tool is necessary to activate new validators`,
 	Example: "lukso network validator deposit_alt",
 	Run: func(cmd *cobra.Command, args []string) {
-		nodeConf, err := network.GetLoadedNodeConfigs()
-		if err != nil {
-			cobra.CompErrorln(err.Error())
-			return
-		}
+		nodeConf := network.MustGetNodeConfig()
 		valSecrets := nodeConf.GetValSecrets()
 		if valSecrets == nil {
 			cobra.CompErrorln("no validator credential is presented")