+++ src/constants.go
@@ -1,6 +1,0 @@
-package src
- 
-// Keys that are used in Viper
-const (
-	ViperKeyNetworkName = "NETWORK_NAME"
-)